style: move security directory
