+++ backend/src/main/java/com/hjjang/backend/domain/user/controller/UserAuthController.java
@@ -4,9 +4,9 @@ import com.hjjang.backend.domain.user.entity.RoleType;
 import com.hjjang.backend.domain.user.entity.UserRefreshToken;
 import com.hjjang.backend.domain.user.repository.UserRefreshTokenRepository;
 import com.hjjang.backend.domain.user.service.UserAuthService;
-import com.hjjang.backend.global.config.properties.AuthProperties;
-import com.hjjang.backend.global.config.security.token.AuthToken;
-import com.hjjang.backend.global.config.security.token.AuthTokenProvider;
+import com.hjjang.backend.global.config.security.properties.AuthProperties;
+import com.hjjang.backend.global.security.token.AuthToken;
+import com.hjjang.backend.global.security.token.AuthTokenProvider;
 import com.hjjang.backend.global.dto.ApiResponse;
 import com.hjjang.backend.global.util.CookieUtil;
 import com.hjjang.backend.global.util.HeaderUtil;
