+++ backend/src/main/java/com/hjjang/backend/global/security/filter/TokenAuthenticationFilter.java
@@ -1,22 +1,19 @@
 package com.hjjang.backend.global.security.filter;
 
-import java.io.IOException;
+import com.hjjang.backend.global.security.token.AuthToken;
+import com.hjjang.backend.global.security.token.AuthTokenProvider;
+import com.hjjang.backend.global.util.HeaderUtil;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.filter.OncePerRequestFilter;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import com.hjjang.backend.global.security.token.AuthToken;
-import com.hjjang.backend.global.security.token.AuthTokenProvider;
-import com.hjjang.backend.global.util.HeaderUtil;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
 
 @Slf4j
 @RequiredArgsConstructor
